+++ test/test_api/test_signatures.py
@@ -70,5 +70,5 @@ def test_param_kind_and_name(code, index, param_code, kind, Script, skip_python2
 
 
 def test_staticmethod(Script):
-    s, = Script('staticmethod(').call_signatures()
+    s, = Script('staticmethod(').find_signatures()
     assert s.to_string() == 'staticmethod(f: Callable[..., Any])'