+++ test/test_inference/test_docstring.py
@@ -422,5 +422,5 @@ def test_basic_str_init_signature(Script, disable_typeshed):
         class Foo(str):
             pass
         Foo(''')
-    c, = Script(code).call_signatures()
+    c, = Script(code).find_signatures()
     assert c.name == 'Foo'